fixes + soft delete

+++ backend/src/main/java/com/cactusshop/backend/dto/OrderStatusResponseDTO.java
@@ -1,5 +1,7 @@
 package com.cactusshop.backend.dto;
 
+import java.math.BigDecimal;
+
 // Răspunsul public la verificarea unei comenzi — expune DOAR ce are nevoie
 // clientul să vadă, nu toată entitatea Order.
-public record OrderStatusResponseDTO(Long id, String status, String purchasedItems, double totalPrice) {}
+public record OrderStatusResponseDTO(Long id, String status, String purchasedItems, BigDecimal totalPrice) {}
